requirements.txt
pytorch-lightning==1.9.*
tqdm
numpy
scipy
seaborn
tensorboard
wandb
xvfbwrapper
scikit-learn
matplotlib
opencv-python
PyYAML
gymnasium
networkx
mne
